requirements_dev.txt
camelot-py==0.9.0
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
et-xmlfile==1.1.0
lxml==5.2.1
multipledispatch==1.0.0
numpy==1.26.4
opencv-python==4.9.0.80
openpyxl==3.1.2
packaging==24.0
pandas==2.2.2
pdfminer.six==20231228
pillow==10.3.0
pycparser==2.22
PyMuPDF==1.24.2
PyMuPDFb==1.24.1
PyPDF2==2.12.1
pytesseract==0.3.10
python-dateutil==2.9.0.post0
python-docx==1.1.0
pytz==2024.1
six==1.16.0
typing_extensions==4.11.0
tzdata==2024.1
twine==5.0.0
